Add MemberExpressionExtension
